refactor code to be easier to match with the code snippets in the book
