+++ src/main/java/io/reflectoring/buckpal/account/adapter/in/web/SendMoneyController.java
@@ -1,0 +1,34 @@
+package io.reflectoring.buckpal.account.adapter.in.web;
+
+import io.reflectoring.buckpal.account.application.port.in.SendMoneyUseCase;
+import io.reflectoring.buckpal.account.application.port.in.SendMoneyCommand;
+import io.reflectoring.buckpal.common.WebAdapter;
+import io.reflectoring.buckpal.account.domain.Account.AccountId;
+import io.reflectoring.buckpal.account.domain.Money;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@WebAdapter
+@RestController
+@RequiredArgsConstructor
+class SendMoneyController {
+
+	private final SendMoneyUseCase sendMoneyUseCase;
+
+	@PostMapping(path = "/accounts/send/{sourceAccountId}/{targetAccountId}/{amount}")
+	void sendMoney(
+			@PathVariable("sourceAccountId") Long sourceAccountId,
+			@PathVariable("targetAccountId") Long targetAccountId,
+			@PathVariable("amount") Long amount) {
+
+		SendMoneyCommand command = new SendMoneyCommand(
+				new AccountId(sourceAccountId),
+				new AccountId(targetAccountId),
+				Money.of(amount));
+
+		sendMoneyUseCase.sendMoney(command);
+	}
+
+}
