+++ src/main/java/io/reflectoring/buckpal/account/application/port/in/GetAccountBalanceQuery.java
@@ -1,0 +1,10 @@
+package io.reflectoring.buckpal.account.application.port.in;
+
+import io.reflectoring.buckpal.account.domain.Account.AccountId;
+import io.reflectoring.buckpal.account.domain.Money;
+
+public interface GetAccountBalanceQuery {
+
+	Money getAccountBalance(AccountId accountId);
+
+}
