+++ src/main/java/io/reflectoring/buckpal/account/application/port/in/SendMoneyUseCase.java
@@ -1,0 +1,11 @@
+package io.reflectoring.buckpal.account.application.port.in;
+
+import io.reflectoring.buckpal.account.domain.Account.AccountId;
+import io.reflectoring.buckpal.account.domain.Money;
+import io.reflectoring.buckpal.common.SelfValidating;
+
+public interface SendMoneyUseCase {
+
+	boolean sendMoney(SendMoneyCommand command);
+
+}
