+++ src/main/java/io/reflectoring/buckpal/account/application/service/NoOpAccountLock.java
@@ -1,0 +1,20 @@
+package io.reflectoring.buckpal.account.application.service;
+
+import io.reflectoring.buckpal.account.application.port.out.AccountLock;
+import io.reflectoring.buckpal.account.domain.Account.AccountId;
+import org.springframework.stereotype.Component;
+
+@Component
+class NoOpAccountLock implements AccountLock {
+
+	@Override
+	public void lockAccount(AccountId accountId) {
+		// do nothing
+	}
+
+	@Override
+	public void releaseAccount(AccountId accountId) {
+		// do nothing
+	}
+
+}
