--- src/main/java/io/reflectoring/buckpal/adapter/persistence/AccountJpaEntity.java
@@ -1,23 +1,0 @@
-package io.reflectoring.buckpal.adapter.persistence;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "account")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-class AccountJpaEntity {
-
-	@Id
-	@GeneratedValue
-	private Long id;
-
-}
