--- src/main/java/io/reflectoring/buckpal/adapter/persistence/AccountMapper.java
@@ -1,60 +1,0 @@
-package io.reflectoring.buckpal.adapter.persistence;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import io.reflectoring.buckpal.domain.Account;
-import io.reflectoring.buckpal.domain.Account.AccountId;
-import io.reflectoring.buckpal.domain.Activity;
-import io.reflectoring.buckpal.domain.Activity.ActivityId;
-import io.reflectoring.buckpal.domain.ActivityWindow;
-import io.reflectoring.buckpal.domain.Money;
-import org.springframework.stereotype.Component;
-
-@Component
-class AccountMapper {
-
-	Account mapToDomainEntity(
-			AccountJpaEntity account,
-			List<ActivityJpaEntity> activities,
-			Long withdrawalBalance,
-			Long depositBalance) {
-
-		Money baselineBalance = Money.subtract(
-				Money.of(depositBalance),
-				Money.of(withdrawalBalance));
-
-		return Account.withId(
-				new AccountId(account.getId()),
-				baselineBalance,
-				mapToActivityWindow(activities));
-
-	}
-
-	ActivityWindow mapToActivityWindow(List<ActivityJpaEntity> activities) {
-		List<Activity> mappedActivities = new ArrayList<>();
-
-		for (ActivityJpaEntity activity : activities) {
-			mappedActivities.add(new Activity(
-					new ActivityId(activity.getId()),
-					new AccountId(activity.getOwnerAccountId()),
-					new AccountId(activity.getSourceAccountId()),
-					new AccountId(activity.getTargetAccountId()),
-					activity.getTimestamp(),
-					Money.of(activity.getAmount())));
-		}
-
-		return new ActivityWindow(mappedActivities);
-	}
-
-	ActivityJpaEntity mapToJpaEntity(Activity activity) {
-		return new ActivityJpaEntity(
-				activity.getId() == null ? null : activity.getId().getValue(),
-				activity.getTimestamp(),
-				activity.getOwnerAccountId().getValue(),
-				activity.getSourceAccountId().getValue(),
-				activity.getTargetAccountId().getValue(),
-				activity.getMoney().getAmount().longValue());
-	}
-
-}
