--- src/main/java/io/reflectoring/buckpal/adapter/persistence/AccountPersistenceAdapter.java
@@ -1,73 +1,0 @@
-package io.reflectoring.buckpal.adapter.persistence;
-
-import javax.persistence.EntityNotFoundException;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-import io.reflectoring.buckpal.application.port.out.LoadAccountPort;
-import io.reflectoring.buckpal.application.port.out.UpdateAccountStatePort;
-import io.reflectoring.buckpal.common.PersistenceAdapter;
-import io.reflectoring.buckpal.domain.Account;
-import io.reflectoring.buckpal.domain.Account.AccountId;
-import io.reflectoring.buckpal.domain.Activity;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@RequiredArgsConstructor
-@PersistenceAdapter
-class AccountPersistenceAdapter implements
-		LoadAccountPort,
-		UpdateAccountStatePort {
-
-	private final AccountRepository accountRepository;
-	private final ActivityRepository activityRepository;
-	private final AccountMapper accountMapper;
-
-	@Override
-	public Account loadAccount(
-					AccountId accountId,
-					LocalDateTime baselineDate) {
-
-		AccountJpaEntity account =
-				accountRepository.findById(accountId.getValue())
-						.orElseThrow(EntityNotFoundException::new);
-
-		List<ActivityJpaEntity> activities =
-				activityRepository.findByOwnerSince(
-						accountId.getValue(),
-						baselineDate);
-
-		Long withdrawalBalance = orZero(activityRepository
-				.getWithdrawalBalanceUntil(
-						accountId.getValue(),
-						baselineDate));
-
-		Long depositBalance = orZero(activityRepository
-				.getDepositBalanceUntil(
-						accountId.getValue(),
-						baselineDate));
-
-		return accountMapper.mapToDomainEntity(
-				account,
-				activities,
-				withdrawalBalance,
-				depositBalance);
-
-	}
-
-	private Long orZero(Long value){
-		return value == null ? 0L : value;
-	}
-
-
-	@Override
-	public void updateActivities(Account account) {
-		for (Activity activity : account.getActivityWindow().getActivities()) {
-			if (activity.getId() == null) {
-				activityRepository.save(accountMapper.mapToJpaEntity(activity));
-			}
-		}
-	}
-
-}
