--- src/main/java/io/reflectoring/buckpal/adapter/persistence/AccountRepository.java
@@ -1,6 +1,0 @@
-package io.reflectoring.buckpal.adapter.persistence;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface AccountRepository extends JpaRepository<AccountJpaEntity, Long> {
-}
