--- src/main/java/io/reflectoring/buckpal/adapter/persistence/ActivityJpaEntity.java
@@ -1,41 +1,0 @@
-package io.reflectoring.buckpal.adapter.persistence;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import java.time.LocalDateTime;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "activity")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-class ActivityJpaEntity {
-
-	@Id
-	@GeneratedValue
-	private Long id;
-
-	@Column
-	private LocalDateTime timestamp;
-
-	@Column
-	private Long ownerAccountId;
-
-	@Column
-	private Long sourceAccountId;
-
-	@Column
-	private Long targetAccountId;
-
-	@Column
-	private Long amount;
-
-}
