--- src/main/java/io/reflectoring/buckpal/application/port/in/GetAccountBalanceQuery.java
@@ -1,10 +1,0 @@
-package io.reflectoring.buckpal.application.port.in;
-
-import io.reflectoring.buckpal.domain.Account.AccountId;
-import io.reflectoring.buckpal.domain.Money;
-
-public interface GetAccountBalanceQuery {
-
-	Money getAccountBalance(AccountId accountId);
-
-}
