--- src/main/java/io/reflectoring/buckpal/application/service/ThresholdExceededException.java
@@ -1,11 +1,0 @@
-package io.reflectoring.buckpal.application.service;
-
-import io.reflectoring.buckpal.domain.Money;
-
-public class ThresholdExceededException extends RuntimeException {
-
-  public ThresholdExceededException(Money threshold, Money actual) {
-    super(String.format("Maximum threshold for transferring money exceeded: tried to transfer %s but threshold is %s!", actual, threshold));
-  }
-
-}
