--- src/test/java/io/reflectoring/buckpal/adapter/persistence/AccountPersistenceAdapterTest.java
@@ -1,55 +1,0 @@
-package io.reflectoring.buckpal.adapter.persistence;
-
-import java.time.LocalDateTime;
-
-import io.reflectoring.buckpal.domain.Account;
-import io.reflectoring.buckpal.domain.Account.AccountId;
-import io.reflectoring.buckpal.domain.ActivityWindow;
-import io.reflectoring.buckpal.domain.Money;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.test.context.jdbc.Sql;
-import static io.reflectoring.buckpal.common.AccountTestData.*;
-import static io.reflectoring.buckpal.common.ActivityTestData.*;
-import static org.assertj.core.api.Assertions.*;
-
-@DataJpaTest
-@Import({AccountPersistenceAdapter.class, AccountMapper.class})
-class AccountPersistenceAdapterTest {
-
-	@Autowired
-	private AccountPersistenceAdapter adapterUnderTest;
-
-	@Autowired
-	private ActivityRepository activityRepository;
-
-	@Test
-	@Sql("AccountPersistenceAdapterTest.sql")
-	void loadsAccount() {
-		Account account = adapterUnderTest.loadAccount(new AccountId(1L), LocalDateTime.of(2018, 8, 10, 0, 0));
-
-		assertThat(account.getActivityWindow().getActivities()).hasSize(2);
-		assertThat(account.calculateBalance()).isEqualTo(Money.of(500));
-	}
-
-	@Test
-	void updatesActivities() {
-		Account account = defaultAccount()
-				.withBaselineBalance(Money.of(555L))
-				.withActivityWindow(new ActivityWindow(
-						defaultActivity()
-								.withId(null)
-								.withMoney(Money.of(1L)).build()))
-				.build();
-
-		adapterUnderTest.updateActivities(account);
-
-		assertThat(activityRepository.count()).isEqualTo(1);
-
-		ActivityJpaEntity savedActivity = activityRepository.findAll().get(0);
-		assertThat(savedActivity.getAmount()).isEqualTo(1L);
-	}
-
-}
