+++ src/test/java/io/reflectoring/buckpal/common/ActivityTestData.java
@@ -2,10 +2,10 @@ package io.reflectoring.buckpal.common;
 
 import java.time.LocalDateTime;
 
-import io.reflectoring.buckpal.domain.Account.AccountId;
-import io.reflectoring.buckpal.domain.Activity;
-import io.reflectoring.buckpal.domain.Activity.ActivityId;
-import io.reflectoring.buckpal.domain.Money;
+import io.reflectoring.buckpal.account.domain.Account.AccountId;
+import io.reflectoring.buckpal.account.domain.Activity;
+import io.reflectoring.buckpal.account.domain.Activity.ActivityId;
+import io.reflectoring.buckpal.account.domain.Money;
 
 public class ActivityTestData {
